+++ at_client/lib/src/preference/at_client_preference.dart
@@ -29,8 +29,8 @@ class AtClientPreference {
   /// Secret key to encrypt keystore data
   List<int>? keyStoreSecret;
 
-  /// Domain of the root server. Defaults to root.atsign.wtf
-  String rootDomain = 'root.atsign.wtf';
+  /// Domain of the root server. Defaults to root.atsign.org
+  String rootDomain = 'root.atsign.org';
 
   /// Port of the root server. Defaults to 64
   int rootPort = 64;