+++ at_client/lib/src/service/file_transfer_service.dart
@@ -4,7 +4,6 @@ import 'package:archive/archive_io.dart';
 import 'package:at_client/src/stream/file_transfer_object.dart';
 import 'package:at_client/src/util/constants.dart';
 import 'package:http/http.dart' as http;
-import 'package:http/http.dart';
 
 class FileTransferService {
   Future<dynamic> uploadToFileBin(
@@ -34,7 +33,7 @@ class FileTransferService {
         streamedRequest.sink.close();
       });
 
-      StreamedResponse response = await streamedRequest.send();
+      http.StreamedResponse response = await streamedRequest.send();
       return response;
     } catch (e) {
       rethrow;