+++ at_client/lib/src/transformer/response_transformer/get_response_transformer.dart
@@ -3,11 +3,9 @@ import 'dart:async';
 import 'package:at_base2e15/at_base2e15.dart';
 import 'package:at_client/at_client.dart';
 import 'package:at_client/src/decryption_service/decryption_manager.dart';
-import 'package:at_client/src/manager/at_client_manager.dart';
 import 'package:at_client/src/response/default_response_parser.dart';
 import 'package:at_client/src/response/json_utils.dart';
 import 'package:at_client/src/transformer/at_transformer.dart';
-import 'package:at_client/src/util/at_client_util.dart';
 import 'package:at_commons/at_commons.dart';
 
 /// Class responsible for transforming the Get response
@@ -30,21 +28,23 @@ class GetResponseTransformer
           decodedResponse['key'].startsWith('public:'));
     }
 
-    // If data is binary, decode the data
-    if (atValue.metadata != null &&
-        atValue.metadata!.isBinary != null &&
-        atValue.metadata!.isBinary!) {
-      atValue.value = Base2e15.decode(atValue.value);
-    }
-
-    // If data is encrypted, decrypt the data
-    if (atValue.metadata != null &&
-        atValue.metadata!.isEncrypted != null &&
-        atValue.metadata!.isEncrypted!) {
+    // For public and cached public keys, data is not encrypted.
+    // Decrypt the data, for other keys
+    if (!(decodedResponse['key'].startsWith('public:')) &&
+        !(decodedResponse['key'].startsWith('cached:public:'))) {
       var decryptionService = AtKeyDecryptionManager.get(tuple.one,
           AtClientManager.getInstance().atClient.getCurrentAtSign()!);
       atValue.value =
           await decryptionService.decrypt(tuple.one, atValue.value) as String;
+    }
+
+    // After decrypting the data, if data is binary, decode the data
+    // For cached keys, isBinary is not on server-side. Hence getting
+    // isBinary from AtKey.
+    if (tuple.one.metadata != null &&
+        tuple.one.metadata!.isBinary != null &&
+        tuple.one.metadata!.isBinary!) {
+      atValue.value = Base2e15.decode(atValue.value);
     }
     return atValue;
   }