+++ at_functional_test/test/set_encryption_keys.dart
@@ -25,6 +25,11 @@ Future<void> setEncryptionKeys(
       ..metadata = metadata;
     result = await atClient.put(
         atKey, demo_credentials.encryptionPublicKeyMap[atsign]);
+
+    // set self encryption key
+    await atClient
+        .getLocalSecondary()!
+        .putValue(AT_ENCRYPTION_SELF_KEY, demo_credentials.aesKeyMap[atsign]!);
     print(result);
   } on Exception catch (e, trace) {
     print(e.toString());