+++ at_functional_test/test/test_utils.dart
@@ -1,0 +1,15 @@
+import 'package:at_client/at_client.dart';
+
+import 'at_demo_credentials.dart' as demo_credentials;
+
+class TestUtils {
+  static AtClientPreference getPreference(String atsign) {
+    var preference = AtClientPreference();
+    preference.hiveStoragePath = 'test/hive/client';
+    preference.commitLogPath = 'test/hive/client/commit';
+    preference.isLocalStoreRequired = true;
+    preference.privateKey = demo_credentials.pkamPrivateKeyMap[atsign];
+    preference.rootDomain = 'vip.ve.atsign.zone';
+    return preference;
+  }
+}